Move fie code to fie
